+++ src/config.ts
@@ -1,0 +1,6 @@
+import type { AbstractStorage } from "./storage";
+
+export type Config = {
+	strict: boolean;
+	storage: AbstractStorage;
+};