+++ src/repositories/as-associate.ts
@@ -1,0 +1,5 @@
+import { CartRepository } from "./cart";
+import { OrderRepository } from "./order";
+
+export class AsAssociateOrderRepository extends OrderRepository {}
+export class AsAssociateCartRepository extends CartRepository {}