+++ src/repositories/product-selection.ts
@@ -2,17 +2,19 @@ import type {
 	ProductSelection,
 	ProductSelectionChangeNameAction,
 	ProductSelectionDraft,
+	ProductSelectionSetCustomTypeAction,
 	ProductSelectionUpdateAction,
 } from "@commercetools/platform-sdk";
+import type { Config } from "~src/config";
+import { createCustomFields } from "~src/repositories/helpers";
 import { getBaseResourceProperties } from "../helpers";
-import type { AbstractStorage } from "../storage/abstract";
 import type { Writable } from "../types";
 import type { RepositoryContext, UpdateHandlerInterface } from "./abstract";
 import { AbstractResourceRepository, AbstractUpdateHandler } from "./abstract";
 
 export class ProductSelectionRepository extends AbstractResourceRepository<"product-selection"> {
-	constructor(storage: AbstractStorage) {
-		super("product-selection", storage);
+	constructor(config: Config) {
+		super("product-selection", config);
 		this.actions = new ProductSelectionUpdateHandler(this._storage);
 	}
 
@@ -44,5 +46,21 @@ class ProductSelectionUpdateHandler
 		{ name }: ProductSelectionChangeNameAction,
 	) {
 		resource.name = name;
+	}
+
+	setCustomType(
+		context: RepositoryContext,
+		resource: Writable<ProductSelection>,
+		{ type, fields }: ProductSelectionSetCustomTypeAction,
+	) {
+		if (type) {
+			resource.custom = createCustomFields(
+				{ type, fields },
+				context.projectKey,
+				this._storage,
+			);
+		} else {
+			resource.custom = undefined;
+		}
 	}
 }